refactor: major codebase update with security improvements and attribution changes
This commit includes comprehensive updates across the entire project:
- Enhanced GitHub Actions workflows with dynamic security groups and custom SSH ports
- Updated attribution in workflow files to acknowledge collaborative development
- Refactored Rust, C++, CUDA, and Python source files
- Updated example notebooks and documentation
- Added new five-twist example variants
- Simplified Docker build workflow

+++ src/cpp/utility/utility.cu
@@ -1,7 +1,7 @@
 // File: utility.cu
-// Author: Ryoichi Ando (ryoichi.ando@zozo.com)
+// Code: Claude Code and Codex
+// Review: Ryoichi Ando (ryoichi.ando@zozo.com)
 // License: Apache v2.0
-
 
 #include "dispatcher.hpp"
 #include "utility.hpp"
@@ -21,7 +21,7 @@ __device__ Vec3f compute_vertex_normal(const DataSet &data,
             const Vec3f &z0 = vertex[face[0]];
             const Vec3f &z1 = vertex[face[1]];
             const Vec3f &z2 = vertex[face[2]];
-            normal += (z1 - z0).cross((z2 - z0).cast<float>());
+            normal += (z1 - z0).cross(z2 - z0);
         }
         if (normal.squaredNorm()) {
             normal.normalize();
@@ -233,12 +233,12 @@ void compute_svd(DataSet data, Vec<Vec3f> curr, Vec<Svd3x2> svd,
     auto svd_data = svd.data;
     auto inv_rest2x2 = data.inv_rest2x2.data;
     DISPATCH_START(shell_face_count)
-    [mesh_face, curr_data, svd_data, inv_rest2x2] __device__(unsigned i) mutable {
+    [mesh_face, curr_data, svd_data,
+     inv_rest2x2] __device__(unsigned i) mutable {
         Vec3u face = mesh_face[i];
         Mat3x3f x;
         x << curr_data[face[0]], curr_data[face[1]], curr_data[face[2]];
-        const Mat3x2f F =
-            utility::compute_deformation_grad(x, inv_rest2x2[i]);
+        const Mat3x2f F = utility::compute_deformation_grad(x, inv_rest2x2[i]);
         svd_data[i] = utility::svd3x2(F);
     } DISPATCH_END;
 }
